Add Merge PDF option in View Files
Fixed #440
