+++ app/src/main/java/swati4star/createpdf/fragment/ViewFilesFragment.java
@@ -32,8 +32,6 @@ import android.widget.LinearLayout;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
 
-import com.afollestad.materialdialogs.MaterialDialog;
-
 import java.io.File;
 import java.util.ArrayList;
 import java.util.Objects;
@@ -44,35 +42,26 @@ import butterknife.OnClick;
 import swati4star.createpdf.R;
 import swati4star.createpdf.activity.MainActivity;
 import swati4star.createpdf.adapter.ViewFilesAdapter;
-import swati4star.createpdf.database.DatabaseHelper;
 import swati4star.createpdf.interfaces.EmptyStateChangeListener;
 import swati4star.createpdf.interfaces.ItemSelectedListener;
-import swati4star.createpdf.interfaces.MergeFilesListener;
 import swati4star.createpdf.util.DirectoryUtils;
 import swati4star.createpdf.util.FileSortUtils;
-import swati4star.createpdf.util.FileUtils;
-import swati4star.createpdf.util.MergePdf;
+import swati4star.createpdf.util.MergeHelper;
 import swati4star.createpdf.util.MoveFilesToDirectory;
 import swati4star.createpdf.util.PopulateList;
-import swati4star.createpdf.util.StringUtils;
 import swati4star.createpdf.util.ViewFilesDividerItemDecoration;
 
 import static swati4star.createpdf.util.Constants.BUNDLE_DATA;
 import static swati4star.createpdf.util.Constants.SORTING_INDEX;
-import static swati4star.createpdf.util.Constants.STORAGE_LOCATION;
 import static swati4star.createpdf.util.Constants.appName;
-import static swati4star.createpdf.util.DialogUtils.createAnimationDialog;
-import static swati4star.createpdf.util.DialogUtils.createOverwriteDialog;
 import static swati4star.createpdf.util.DialogUtils.showFilesInfoDialog;
 import static swati4star.createpdf.util.FileSortUtils.NAME_INDEX;
-import static swati4star.createpdf.util.StringUtils.getDefaultStorageLocation;
-import static swati4star.createpdf.util.StringUtils.getSnackbarwithAction;
 import static swati4star.createpdf.util.StringUtils.showSnackbar;
 
 public class ViewFilesFragment extends Fragment
         implements SwipeRefreshLayout.OnRefreshListener,
         EmptyStateChangeListener,
-        ItemSelectedListener, MergeFilesListener {
+        ItemSelectedListener {
 
     // Directory operations constants
     public static final int NEW_DIR = 1;
@@ -106,12 +95,8 @@ public class ViewFilesFragment extends Fragment
     private AlertDialog.Builder mAlertDialogBuilder;
 
 
-    private FileUtils mFileUtils;
-    private String mHomePath;
-    private boolean mPasswordProtected = false;
-    private String mPassword;
     private int mCountFiles;
-    private MaterialDialog mMaterialDialog;
+    private MergeHelper mMergeHelper;
 
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
@@ -132,9 +117,6 @@ public class ViewFilesFragment extends Fragment
         mViewFilesListRecyclerView.addItemDecoration(new ViewFilesDividerItemDecoration(root.getContext()));
         mSwipeView.setOnRefreshListener(this);
 
-        mHomePath = PreferenceManager.getDefaultSharedPreferences(mActivity)
-                .getString(STORAGE_LOCATION,
-                        getDefaultStorageLocation());
         int dialogId;
         if (getArguments() != null) {
             dialogId = getArguments().getInt(BUNDLE_DATA);
@@ -142,6 +124,7 @@ public class ViewFilesFragment extends Fragment
         }
 
         checkIfListEmpty();
+        mMergeHelper = new MergeHelper(mActivity, mViewFilesAdapter);
         return root;
     }
 
@@ -221,61 +204,13 @@ public class ViewFilesFragment extends Fragment
                 break;
             case R.id.item_merge:
                 if (mViewFilesAdapter.getItemCount() > 1) {
-                    showSnackbar(mActivity, "merged");
-                    mergeFiles();
+                    mMergeHelper.mergeFiles();
                 }
                 break;
         }
         return true;
     }
 
-    /**
-     * Merge the selected PDFs into a single PDF
-     * Show a dialog to allow user to enter file name
-     */
-    void mergeFiles() {
-        String[] pdfpaths =  mViewFilesAdapter.getSelectedFilePath().toArray(new String[0]);
-        new MaterialDialog.Builder(mActivity)
-                .title(R.string.creating_pdf)
-                .content(R.string.enter_file_name)
-                .input(getString(R.string.example), null, (dialog, input) -> {
-                    if (StringUtils.isEmpty(input)) {
-                        showSnackbar(mActivity, R.string.snackbar_name_not_blank);
-                    } else {
-                        if (!mFileUtils.isFileExist(input + getString(R.string.pdf_ext))) {
-                            new MergePdf(input.toString(), mHomePath, mPasswordProtected,
-                                    mPassword, this).execute(pdfpaths);
-                        } else {
-                            MaterialDialog.Builder builder = createOverwriteDialog(mActivity);
-                            builder.onPositive((dialog12, which) -> new MergePdf(input.toString(),
-                                    mHomePath, mPasswordProtected, mPassword,
-                                    this).execute(pdfpaths))
-                                    .onNegative((dialog1, which) -> mergeFiles()).show();
-                        }
-                    }
-                })
-                .show();
-    }
-
-    @Override
-    public void resetValues(boolean isPDFMerged, String path) {
-        mMaterialDialog.dismiss();
-        if (isPDFMerged) {
-            getSnackbarwithAction(mActivity, R.string.pdf_merged)
-                    .setAction(R.string.snackbar_viewAction, v -> mFileUtils.openFile(path)).show();
-            new DatabaseHelper(mActivity).insertRecord(path,
-                    mActivity.getString(R.string.created));
-        } else
-            showSnackbar(mActivity, R.string.pdf_merge_error);
-
-        mViewFilesAdapter.updateDataset();
-    }
-
-    @Override
-    public void mergeStarted() {
-        mMaterialDialog = createAnimationDialog(mActivity);
-        mMaterialDialog.show();
-    }
 
     /**
      * Moves files from one directory to another
@@ -550,7 +485,6 @@ public class ViewFilesFragment extends Fragment
     public void onAttach(Context context) {
         super.onAttach(context);
         mActivity = (Activity) context;
-        mFileUtils = new FileUtils(mActivity);
         mDirectoryUtils = new DirectoryUtils(mActivity);
     }
