+++ app/src/main/java/swati4star/createpdf/util/MergeHelper.java
@@ -1,0 +1,80 @@
+package swati4star.createpdf.util;
+
+import android.app.Activity;
+import android.content.Context;
+import android.preference.PreferenceManager;
+import com.afollestad.materialdialogs.MaterialDialog;
+import swati4star.createpdf.R;
+import swati4star.createpdf.adapter.ViewFilesAdapter;
+import swati4star.createpdf.database.DatabaseHelper;
+import swati4star.createpdf.interfaces.MergeFilesListener;
+
+import static swati4star.createpdf.util.Constants.STORAGE_LOCATION;
+import static swati4star.createpdf.util.DialogUtils.createAnimationDialog;
+import static swati4star.createpdf.util.DialogUtils.createOverwriteDialog;
+import static swati4star.createpdf.util.StringUtils.getDefaultStorageLocation;
+import static swati4star.createpdf.util.StringUtils.getSnackbarwithAction;
+import static swati4star.createpdf.util.StringUtils.showSnackbar;
+
+public class MergeHelper implements MergeFilesListener {
+    private MaterialDialog mMaterialDialog;
+    private Activity mActivity;
+    private FileUtils mFileUtils;
+    private boolean mPasswordProtected = false;
+    private String mPassword;
+    private String mHomePath;
+    private Context mContext;
+    private ViewFilesAdapter mViewFilesAdapter;
+
+    public MergeHelper(Activity activity, ViewFilesAdapter viewFilesAdapter) {
+        mActivity = activity;
+        mFileUtils = new FileUtils(mActivity);
+        mHomePath = PreferenceManager.getDefaultSharedPreferences(mActivity)
+                .getString(STORAGE_LOCATION,
+                        getDefaultStorageLocation());
+        mContext = mActivity;
+        mViewFilesAdapter = viewFilesAdapter;
+    }
+
+    public void mergeFiles() {
+        String[] pdfpaths = mViewFilesAdapter.getSelectedFilePath().toArray(new String[0]);
+        new MaterialDialog.Builder(mActivity)
+                .title(R.string.creating_pdf)
+                .content(R.string.enter_file_name)
+                .input(mContext.getResources().getString(R.string.example), null, (dialog, input) -> {
+                    if (StringUtils.isEmpty(input)) {
+                        showSnackbar(mActivity, R.string.snackbar_name_not_blank);
+                    } else {
+                        if (!mFileUtils.isFileExist(input + mContext.getResources().getString(R.string.pdf_ext))) {
+                            new MergePdf(input.toString(), mHomePath, mPasswordProtected,
+                                    mPassword, this).execute(pdfpaths);
+                        } else {
+                            MaterialDialog.Builder builder = createOverwriteDialog(mActivity);
+                            builder.onPositive((dialog12, which) -> new MergePdf(input.toString(),
+                                    mHomePath, mPasswordProtected, mPassword,
+                                    this).execute(pdfpaths))
+                                    .onNegative((dialog1, which) -> mergeFiles()).show();
+                        }
+                    }
+                })
+                .show();
+    }
+    @Override
+    public void resetValues(boolean isPDFMerged, String path) {
+        mMaterialDialog.dismiss();
+        if (isPDFMerged) {
+            getSnackbarwithAction(mActivity, R.string.pdf_merged)
+                    .setAction(R.string.snackbar_viewAction, v -> mFileUtils.openFile(path)).show();
+            new DatabaseHelper(mActivity).insertRecord(path,
+                    mActivity.getString(R.string.created));
+        } else
+            showSnackbar(mActivity, R.string.pdf_merge_error);
+        mViewFilesAdapter.updateDataset();
+    }
+
+    @Override
+    public void mergeStarted() {
+        mMaterialDialog = createAnimationDialog(mActivity);
+        mMaterialDialog.show();
+    }
+}
